fix error produced by merging
